// src/app/api/models/register-company-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegisterCompanyDto {
  adminEmail?: string | null;
  city?: string | null;
  companyName?: string | null;
  country?: string | null;
  firstName?: string | null;
  houseNumber?: number | null;
  lastName?: string | null;
  password?: string | null;
  postalCode?: number | null;
  street?: string | null;
}